commit 4 - Menambah fitur
Menambah fitur kemudahan pengguna

+++ hapus.php
@@ -1,4 +1,9 @@
 <?php
+session_start();
+if (!isset($_SESSION['user_id'])) {
+    header('Location: login.php');
+    exit;
+}
 include 'koneksi.php';
 $id = $_GET['id'];
 mysqli_query($conn, "DELETE FROM penghuni WHERE id=$id");
